Building a First Custom Component
A component in javascript is just a function. The convention is that the name of the function is the same as the file but without .js of course. In the end, the function is exported as default because it has only this function in the file. The component was imported in the App.js and used as a html element. The name of the component when used should be the same as the name when it was imported.