Adding Basic CSS Styling
In this commit, the file containing the ExpenseItem component's css was created in the same folder with the same name, just differentiating it with the .css ending. The ExpenseItem.css file was injected inside the ExpenseItem.js component. To name a class in html within js you must use className and not just class as in normal html.