Passing Data via PROPS
Props is the object which holds all the values we get for the attributes on our custom element in the App function located in App.js. Props is a super important concept because it allows you to make your components reusable and it allows you to pass data from another component to this component.

In this example, props were used to pass data from the parent component to the ExpenseItem child component. The parent component is passing three props to ExpenseItem: title, amount and date. Each prop contains a specific value extracted from the expenses0 object.