Working with State
The useState hook is used to create and manage state in a functional component in React. In this example code, the useState hook is used to create a state variable called title and a function called setTitle to update the value of title. The initial value of title is set to props.title. The clickHandler function uses the setTitle function to update the value of title to 'Updated' when a button is clicked. This new value of title is then rendered in the component.