Child-to-Parent Component Communication another example
In this commit a props called onAddExpense was created in the App.js to allow communication with the NewExpense.js. In the NewExpense.js a props was passed in the NewExpense as parameter and the props was called in the end of the NewExpense function code passing the stored expenseData information to the App.js. To prove the data is coming from App.js a console.log was created in line 30 and that is the one that appears in the console.