Outputting Conditional Content
In this commit an filter method for checking the expenses by year was implemented in the Expenses.js. A filteredExpenses function was created to check if the expense date is equal to filteredYear. As expense.data is an object and filteredYear is a string. The expense data was converted to a string for the === check be possible. After expensesContent was created and stored outside the JSX using only If statements in a modern way. Then expensesContent was called in JSX.